% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcftocsv.R
\name{VCF_to_CSV}
\alias{VCF_to_CSV}
\title{VCF to CSV conversion}
\usage{
VCF_to_CSV(input_file)
}
\description{
Extract genotypes from VCF files and convert to CSV
Input can be a VCF file or excel file
}
\examples{
input_file <- VCF_to_CSV("my_vcf_file.vcf");
}
